zenquotes/requirements.txt
requests
psycopg2-binary
python-dotenv
email-validator